requirements.txt
django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
whitenoise
gunicorn
psycopg2-binary
python-dotenv
dj-database-url 